Code structure change
